+++ modules/default/weather/weather.js
@@ -11,8 +11,10 @@ Module.register("weather", {
 	defaults: {
 		weatherProvider: "openweathermap",
 		roundTemp: false,
-		type: "current", //current, forecast
+		type: "current", // current, forecast, daily (equivalent to forecast), hourly (only with OpenWeatherMap /onecall endpoint)
 
+		lat: 0,
+		lon: 0,
 		location: false,
 		locationID: false,
 		units: config.units,
@@ -36,6 +38,7 @@ Module.register("weather", {
 		showIndoorTemperature: false,
 		showIndoorHumidity: false,
 		maxNumberOfDays: 5,
+		maxEntries: 5,
 		fade: true,
 		fadePoint: 0.25, // Start on 1/4th of the list.
 
@@ -125,7 +128,17 @@ Module.register("weather", {
 
 	// Select the template depending on the display type.
 	getTemplate: function () {
-		return `${this.config.type.toLowerCase()}.njk`;
+		switch (this.config.type.toLowerCase()) {
+			case "current":
+				return `current.njk`;
+			case "hourly":
+				return `hourly.njk`;
+			case "daily":
+			case "forecast":
+				return `forecast.njk`;
+			default:
+				return `${this.config.type.toLowerCase()}.njk`;
+		}
 	},
 
 	// Add all the data to the template.
@@ -134,6 +147,7 @@ Module.register("weather", {
 			config: this.config,
 			current: this.weatherProvider.currentWeather(),
 			forecast: this.weatherProvider.weatherForecast(),
+			weatherData: this.weatherProvider.weatherData(),
 			indoor: {
 				humidity: this.indoorHumidity,
 				temperature: this.indoorTemperature
@@ -155,7 +169,9 @@ Module.register("weather", {
 		}
 
 		setTimeout(() => {
-			if (this.config.type === "forecast") {
+			if (this.config.weatherEndpoint === "/onecall") {
+				this.weatherProvider.fetchWeatherData();
+			} else if (this.config.type === "forecast") {
 				this.weatherProvider.fetchWeatherForecast();
 			} else {
 				this.weatherProvider.fetchCurrentWeather();
@@ -207,7 +223,7 @@ Module.register("weather", {
 						}
 					}
 				} else if (type === "precip") {
-					if (isNaN(value) || value === 0 || value.toFixed(2) === "0.00") {
+					if (value === null || isNaN(value) || value === 0 || value.toFixed(2) === "0.00") {
 						value = "";
 					} else {
 						if (this.config.weatherProvider === "ukmetoffice" || this.config.weatherProvider === "ukmetofficedatahub") {
@@ -242,6 +258,13 @@ Module.register("weather", {
 			"calcNumSteps",
 			function (forecast) {
 				return Math.min(forecast.length, this.config.maxNumberOfDays);
+			}.bind(this)
+		);
+
+		this.nunjucksEnvironment().addFilter(
+			"calcNumEntries",
+			function (dataArray) {
+				return Math.min(dataArray.length, this.config.maxEntries);
 			}.bind(this)
 		);